Rename SSD files as suggested in code review
