+++ luminoth/models/ssd/proposal.py
